Fixed auto-exposure lock
Was not working on S3 (too bad I use two of them for this project).
Fixed it by adding a "lock exposure" button on main screen; looks like
it's impossible to lock auto-exposure if camera is not working. Thanks
to "Open Camera" I found that.
Works on CM11 so far.